read and write into files

+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace UsingFiles
 {
@@ -16,17 +17,17 @@ namespace UsingFiles
             //string docPath = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
             //string userPath = $"C:{Path.DirectorySeparatorChar}User";
             var salesfiles = FindFiles(storesDirectory);
+            var salesTotal = CalculateSalesTotal(salesfiles);
 
-            File.WriteAllText(Path.Combine(salesTotalsDir,"totals.txt"), String.Empty);
+            if(Directory.Exists(Path.Combine(salesTotalsDir,"totals.txt")) == true)
+            {
+                File.WriteAllText(Path.Combine(salesTotalsDir,"totals.txt"), String.Empty);
+            }
+
+            File.AppendAllText(Path.Combine(salesTotalsDir,"totals.txt"), $"{salesTotal}{Environment.NewLine}");
             Console.WriteLine(currentPath);
  
-            //string fileName = $"stores{Path.DirectorySeparatorChar}201{Path.DirectorySeparatorChar}sales{Path.DirectorySeparatorChar}sales.json";
-//
-            //FileInfo info = new FileInfo(fileName);
-//
-            //Console.WriteLine($"Full Name: {info.FullName}{Environment.NewLine}Directory: {info.Directory}{Environment.NewLine}Extension: {info.Extension}{Environment.NewLine}Create Date: {info.CreationTime}"); // And many more
         }
-
         static IEnumerable<string> FindFiles(string folderName)
         {
             List<string> salesFiles = new List<string>();
@@ -45,6 +46,26 @@ namespace UsingFiles
     
             return salesFiles;
         }
+        static double CalculateSalesTotal(IEnumerable<string> salesFiles)
+        {
+            double salesTotal = 0;
+            //read files loop
+            foreach(var file in salesFiles)
+            {
+                string salesJson = File.ReadAllText(file);
+                //Parse the contents as JSON
+                SalesData data = JsonConvert.DeserializeObject<SalesData>(salesJson);
+                //Add the amount found in the total field to the salesTotal varaible
+                salesTotal += data.Total;
+            }
+            return salesTotal;
+        }
+
+        class SalesData
+        {
+            public double Total{ get; set;}
+        }
     }
+
 
 }
